adding importance,RR and BRDF

+++ RayCaster_2001/MaterialProperties.cpp
@@ -31,10 +31,10 @@ bool MaterialProperties::RussianRoulette()
 	return false;
 
 }
- double * MaterialProperties::BRDF()
+ float * MaterialProperties::BRDF()
 {
 	srand(time(NULL)); // seed my random according to the time ( can be done better by using header <random>)
-	double * angles = new double[2];
+	float * angles = new float[2];
 	angles[0] = (rand() / (RAND_MAX + 1.)) * 2 * M_PI; // generate a random number (0-1) and multiply by 2pi. ~azimuth
 	angles[1] = (rand() / (RAND_MAX + 1.))*(M_PI / 2); //incline
