+++ RayCaster_2001/Ray.h
@@ -15,7 +15,7 @@ public:
 	ColorDbl currentColor = ColorDbl(0,0,0);
 	float currentDistance = MAXILON;
 	Vertex intersectionPoint;
-	float importance;
+	double importance;
 
 	Ray *parent = nullptr;
 	Ray *reflectedRay = nullptr;
@@ -25,7 +25,7 @@ public:
 	ColorDbl surfaceCollision(Scene *scene, int num);
 
 	Ray();
-	Ray(Vertex s, Vertex e, float in);
+	Ray(Vertex s, Vertex e, double in);
 	~Ray();
 };
